fix admin and role user

+++ app/Http/Controllers/AdminController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Validation\Rule;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Product;
 use App\Models\Blog;
 use App\Models\Category;
+use App\Models\User;
 
 class AdminController extends Controller
 {
@@ -16,6 +18,8 @@ class AdminController extends Controller
         if (Auth::check()) {
             $name = Auth::user()->name;
             return view('admin.dashboard', compact('name'));
+        } else {
+            return redirect()->intended('admin/login');
         }
     }
 
@@ -32,16 +36,19 @@ class AdminController extends Controller
 
     public function editProduct($id)
     {
-        $name = Auth::user()->name;
-        $category = Category::all();
         if (Auth::check()) {
-            $pEdit = Product::where('product_id', $id)->first();
+            $name = Auth::user()->name;
+            $category = Category::all();
+            if(Auth::user()->role < 2 ) {
+                $pEdit = Product::where('product_id', $id)->first();
 
-            return view('admin.product.edit', compact('pEdit', 'name', 'category'));
+                return view('admin.product.edit', compact('pEdit', 'name', 'category'));
+            } return redirect()->intended('admin/product')->with('error', 'Bạn không có quyền chỉnh sửa sản phẩm này');
         } else {
             return redirect()->intended('admin');
         }
     }
+
     public function updateProduct($id, Request $request)
     {
         if (Auth::check()) {
@@ -98,7 +105,9 @@ class AdminController extends Controller
         $category = Category::all();
         if (Auth::check()) {
 
-            return view('admin.product.insert', compact('name', 'category'));
+            if(Auth::user()->role < 2 ) {
+                return view('admin.product.insert', compact('name', 'category'));
+            } return redirect()->intended('admin/product')->with('error', 'Bạn không có quyền thêm mới sản phẩm');
         } else {
             return redirect()->intended('admin');
         }
@@ -150,15 +159,19 @@ class AdminController extends Controller
     public function deleteProduct($id)
     {
         if (Auth::check()) {
-            $file = Product::where('product_id', $id)->first();
-            $fileName = $file->image;
-            if (file_exists(public_path('/frontend/img/product/' . $fileName))) {
-                unlink(public_path('/frontend/img/product/' . $fileName));
-            }
+            if(Auth::user()->role < 2 ) {
+                $file = Product::where('product_id', $id)->first();
+                $fileName = $file->image;
+                if (file_exists(public_path('/frontend/img/product/' . $fileName))) {
+                    unlink(public_path('/frontend/img/product/' . $fileName));
+                }
 
-            Product::where('product_id', $id)
-                ->delete();
-            return redirect()->intended('admin/product')->with('success', 'Xóa thành công !');
+                Product::where('product_id', $id)
+                    ->delete();
+                return redirect()->intended('admin/product')->with('success', 'Xóa thành công !');
+            } else {
+                return redirect()->intended('admin/product')->with('error', 'Bạn không có quyền xóa sản phẩm này');
+            }
         } else {
             return redirect()->intended('admin');
         }
@@ -178,9 +191,13 @@ class AdminController extends Controller
     public function editBlog($id)
     {
         if (Auth::check()) {
-            $data = Blog::where('blog_id', $id)->first();
-            $name = Auth::user()->name;
-            return view('admin.blog.edit', compact('data', 'name'));
+            if(Auth::user()->role < 2 ) {
+                $name = Auth::user()->name;
+                $data = Blog::where('blog_id', $id)->first();
+                return view('admin.blog.edit', compact('name', 'data'));
+            } else {
+                return redirect()->intended('admin/blog')->with('error', 'Bạn không có quyền chỉnh sửa bài viết này');
+            }
         } else {
             return redirect()->intended('admin');
         }
@@ -202,7 +219,7 @@ class AdminController extends Controller
                 'mimes' => 'Chỉ nhận các file có đuôi .jpeg, .jpg, .png, .gif',
             ]);
 
-            if($request->has('image')){
+            if ($request->has('image')) {
                 $file = $request->image;
                 $fileName = $file->getClientOriginalName();
 
@@ -237,8 +254,12 @@ class AdminController extends Controller
     public function viewInsertBlog()
     {
         if (Auth::check()) {
-            $name = Auth::user()->name;
-            return view('admin.blog.insert', compact('name'));
+            if(Auth::user()->role < 2 ) {
+                $name = Auth::user()->name;
+                return view('admin.blog.insert', compact('name'));
+            } else {
+                return redirect()->intended('admin/blog')->with('error', 'Bạn không có quyền thêm mới bài viết');
+            }
         } else {
             return redirect()->intended('admin');
         }
@@ -261,7 +282,7 @@ class AdminController extends Controller
                 'mimes' => 'Chỉ nhận các file có đuôi .jpeg, .jpg, .png, .gif',
             ]);
 
-            if($request->has('image')){
+            if ($request->has('image')) {
                 $file = $request->image;
                 $fileName = $file->getClientOriginalName();
 
@@ -285,20 +306,24 @@ class AdminController extends Controller
     public function deleteBlog($id)
     {
         if (Auth::check()) {
-            $file = Blog::where('blog_id', $id)->first();
-            $fileName = $file->img;
-            if (file_exists(public_path('/frontend/img/blog/' . $fileName))) {
-                unlink(public_path('/frontend/img/blog/' . $fileName));
-            }
-            Blog::where('blog_id', $id)
-                ->delete();
+            if(Auth::user()->role < 2 ) {
+                $oldFile = Blog::where('blog_id', $id)->first();
+                $oldFileName = $oldFile->img;
 
-            return redirect()->intended('admin/blog')->with('success', 'Xóa thành công !');
+                if (file_exists(public_path('/frontend/img/blog/' . $oldFileName))) {
+                    unlink(public_path('/frontend/img/blog/' . $oldFileName));
+                }
+
+                Blog::where('blog_id', $id)->delete();
+
+                return redirect()->intended('admin/blog')->with('success', 'Xóa thành công !');
+            } else {
+                return redirect()->intended('admin/blog')->with('error', 'Bạn không có quyền xóa bài viết này');
+            }
         } else {
             return redirect()->intended('admin');
         }
     }
-
 
     public function category()
     {
@@ -313,9 +338,15 @@ class AdminController extends Controller
     public function editCategory($id)
     {
         if (Auth::check()) {
-            $data = Category::where('category_id', $id)->first();
+            $admin_role = Auth::user()->role;
             $name = Auth::user()->name;
-            return view('admin.category.edit', compact('name', 'data'));
+            $data = Category::where('id', $id)->first();
+
+            if($admin_role < 2 ) {
+                return view('admin.category.edit', compact('name', 'data'));
+            } else {
+                return redirect()->intended('admin/category')->with('error', 'Bạn không có quyền sửa danh mục này');
+            }
         } else
             return redirect()->intended('admin');
     }
@@ -324,7 +355,7 @@ class AdminController extends Controller
     {
         if (Auth::check()) {
             $request->validate([
-                'name' => 'required|unique:category,category_name'
+                'name' => 'required|unique:category,category_name,'.$id,
             ], [
                 'name.required' => 'Không được để trống',
                 'name.unique' => 'Tên danh mục đã tồn tại'
@@ -333,7 +364,7 @@ class AdminController extends Controller
                 'category_name' => $request->name
             ];
 
-            Category::where('category_id', $id)
+            Category::where('id', $id)
                 ->update($newData);
 
             return redirect()->intended('admin/category')->with('success', 'Sửa thành công !');
@@ -343,9 +374,16 @@ class AdminController extends Controller
 
     public function viewInsertCategory()
     {
-        $name = Auth::user()->name;
+        
         if (Auth::check()) {
-            return view('admin.category.insert', compact('name'));
+            $admin_role = Auth::user()->role;
+            $name = Auth::user()->name;
+
+            if($admin_role < 2 ) {
+                return view('admin.category.insert', compact('name'));
+            } else {
+                return redirect()->intended('admin/category')->with('error', 'Bạn không có quyền thêm danh mục');
+            }
         } else
             return redirect()->intended('admin');
     }
@@ -375,15 +413,154 @@ class AdminController extends Controller
     public function deleteCategory($id)
     {
         if (Auth::check()) {
-            $count = Category::join('product', 'product.category_id', '=', 'category.category_id')
-                ->where('category.category_id', $id)->count();
+            $count = Category::join('product', 'product.category_id', '=', 'category.id')
+                ->where('category.id', $id)->count();
 
-            if ($count == 0) {
-                Category::where('category_id', $id)
-                    ->delete();
-                return redirect()->intended('admin/category')->with('success', 'Xóa thành công !');
+            $admin_role = Auth::user()->role;
+            if($admin_role < 2 ) {
+                if ($count == 0) {
+                    Category::where('id', $id)
+                        ->delete();
+                    return redirect()->intended('admin/category')->with('success', 'Xóa thành công !');
+                } else {
+                    return redirect()->intended('admin/category')->with('error', 'Không thể xóa danh mục do đang có sản phẩm !');
+                }
             } else {
-                return redirect()->intended('admin/category')->with('error', 'Không thể xóa danh mục này !');
+                return redirect()->intended('admin/category')->with('error', 'Bạn không có quyền xóa danh mục này');
+            }
+        } else
+            return redirect()->intended('admin');
+    }
+
+    public function user()
+    {
+        if (Auth::check()) {
+            $user = User::paginate(10);
+            $name = Auth::user()->name;
+            return view('admin.user.user', compact('name', 'user'));
+        } else
+            return redirect()->intended('admin');
+    }
+
+    public function viewInsertUser()
+    {
+        $name = Auth::user()->name;
+        if (Auth::check()) {
+            if(Auth::user()->role < 2 ) {
+                return view('admin.user.insert', compact('name'));
+            } else {
+                return redirect()->intended('admin/user')->with('error', 'Bạn không có quyền thêm mới tài khoản quản trị');
+            }
+        } else {
+            return redirect()->intended('admin');
+        }
+    }
+
+    public function insertUser(Request $request)
+    {
+        if (Auth::check()) {
+            $request->validate([
+                'name' => 'required|unique:users|min:6|max:100',
+                'email' => 'required|unique:users|email',
+                'date' => 'required|after_or_equal:today',
+                'password' => 'required|min:6|max:50',
+            ], [
+                'required' => 'Không được bỏ trống bất kỳ trường dữ liệu nào',
+                'email' => 'Vui lòng nhập email chính xác',
+                'name.min' => 'Tên tối thiểu gồm 6 kí tự',
+                'name.max' => 'Tên tối đa là 100 kí tự',
+                'password.min' => 'Mật khẩu tối thiểu gồm 6 kí tự',
+                'password.max' => 'Mật khẩu tối đa là 50 kí tự',
+                'after_or_equal' => 'Ngày tạo phải lớn hơn hoặc bằng ngày hiện tại',
+                'name.unique' => 'Tên đăng nhập đã tồn tại',
+                'email.unique' => 'Email đã tồn tại'
+            ]);
+
+
+            $newData = new User();
+            $newData->name = $request->name;
+            $newData->email = $request->email;
+            $newData->password = bcrypt($request->password);
+            $newData->created_at = $request->date;
+            $newData->role = $request->role;
+            $newData->save();
+
+            return redirect()->intended('admin/user')->with('success', 'Thêm mới thành công !');
+        } else {
+            return redirect()->intended('admin');
+        }
+    }
+
+
+    public function editUser($id)
+    {
+        if (Auth::check()) {
+            $admin_role = Auth::user()->role;
+            if($admin_role < 2 ) {
+                $user = User::where('id', $id)->first();
+                $name = Auth::user()->name;
+                if($admin_role < $user->role) {
+                    return view('admin.user.edit', compact('name', 'user'));
+                } else {
+                    return redirect()->intended('admin/user')->with('error', 'Bạn không có quyền sửa mục này');
+                }
+            } else {
+                return redirect()->intended('admin/user')->with('error', 'Bạn không có quyền chỉnh sửa tài khoản quản trị');
+            }
+        } else
+            return redirect()->intended('admin');
+    }
+
+    public function updateUser($id, Request $request)
+    {
+        if (Auth::check()) {
+            $request->validate([
+                'name' => 'required|unique:users,id,' . $id . '|min:6|max:100',
+                'email' => 'required|unique:users,id,' . $id . '|email',
+                'date' => 'required|after_or_equal:today',
+                'password' => 'required|min:6|max:50',
+            ], [
+                'required' => 'Không được bỏ trống bất kỳ trường dữ liệu nào',
+                'email' => 'Vui lòng nhập email chính xác',
+                'name.min' => 'Tên tối thiểu gồm 6 kí tự',
+                'name.max' => 'Tên tối đa là 100 kí tự',
+                'password.min' => 'Mật khẩu tối thiểu gồm 6 kí tự',
+                'password.max' => 'Mật khẩu tối đa là 50 kí tự',
+                'after_or_equal' => 'Ngày tạo phải lớn hơn hoặc bằng ngày hiện tại',
+                'name.unique' => 'Tên đăng nhập đã tồn tại',
+                'email.unique' => 'Email đã tồn tại'
+            ]);
+            $newData = [
+                'name' => $request->name,
+                'email' => $request->email,
+                'password' => bcrypt($request->password),
+                'created_at' => $request->date,
+                'role' => $request->role,
+            ];
+
+            User::where('id', $id)
+                ->update($newData);
+
+            return redirect()->intended('admin/user')->with('success', 'Sửa thành công !');
+        } else
+            return redirect()->intended('admin');
+    }
+
+
+    public function deleteUser($id)
+    {
+        if (Auth::check()) {
+            $admin_role = Auth::user()->role;
+            if($admin_role < 2 ) {
+                $user = User::where('id', $id)->first();
+                if($admin_role < $user->role) {
+                    $user->delete();
+                    return redirect()->intended('admin/user')->with('success', 'Xóa thành công !');
+                } else {
+                    return redirect()->intended('admin/user')->with('error', 'Bạn không có quyền xóa người dùng này');
+                }
+            } else {
+                return redirect()->intended('admin/user')->with('error', 'Bạn không có quyền xóa người dùng này');
             }
         } else
             return redirect()->intended('admin');
